Add basic support for multiple effectors
