+++ bbx_dsp/src/effector.rs
@@ -1,39 +1,16 @@
-use std::fmt::{Display, Formatter};
-
-use crate::{operation::Operation, process::Process, sample::Sample};
+use crate::{effectors::overdrive::OverdriveEffector, operation::Operation};
 
 /// A type of DSP `Block` that produces an output signal by modifying an input signal.
-pub struct Effector;
+pub enum Effector {
+    Overdrive(),
+}
 
 impl Effector {
-    pub fn new() -> Effector {
-        return Effector {};
-    }
-
     pub fn to_operation(self) -> Operation {
-        return Box::new(self);
-    }
-}
-
-impl Display for Effector {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "Effector")
-    }
-}
-
-impl Process for Effector {
-    #[inline]
-    fn process(&mut self, sample: Option<Sample>) -> Sample {
-        return if let Some(sample_result) = sample {
-            if sample_result > 1.0 {
-                1.0
-            } else if sample_result < -1.0 {
-                -1.0
-            } else {
-                sample_result - (sample_result.powi(3) / 3.0)
-            }
-        } else {
-            0.0
+        let effector = match self {
+            Effector::Overdrive() => OverdriveEffector,
         };
+
+        return Box::new(effector);
     }
 }
